requirements.txt
# Core Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Financial Data
yfinance>=0.2.0
pandas-datareader>=0.10.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Excel Support
openpyxl>=3.0.0

# Trading
ib-insync>=0.9.70

# Telegram Bot
pyTelegramBotAPI>=4.0.0

# Configuration Management
python-dotenv>=0.19.0

# Optional: Technical Analysis (used in Jupyter notebooks)
TA-Lib>=0.4.0
